ft: add in more option fields
