+++ src/dotenv.net/DotEnv.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using dotenv.net.DependencyInjection.Infrastructure;
 
 namespace dotenv.net
 {
