+++ tests/dotenv.net.Tests/DotEnv.Tests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.IO;
 using System.Text;
-using dotenv.net.DependencyInjection.Infrastructure;
 using FluentAssertions;
 using Xunit;
